factorize model for pagination filters
