+++ packages/app-builder/src/models/pagination.ts
@@ -1,0 +1,10 @@
+export type PaginationParams = {
+  next?: true | false;
+  previous?: true | false;
+  offsetId?: string;
+  order?: 'ASC' | 'DESC';
+  sorting?: 'created_at';
+  limit?: number;
+};
+
+export type FiltersWithPagination<T> = T & PaginationParams;
