+++ packages/app-builder/src/repositories/DecisionRepository.ts
@@ -1,31 +1,33 @@
 import { type PaginatedResponse } from '@app-builder/components';
 import { type MarbleApi } from '@app-builder/infra/marble-api';
+import { type FiltersWithPagination } from '@app-builder/models/pagination';
 import { add } from 'date-fns';
 import { type Decision, type Outcome } from 'marble-api';
 import { Temporal } from 'temporal-polyfill';
 
+export type DecisionFilters = {
+  outcome?: Outcome[];
+  triggerObject?: string[];
+  dateRange?:
+    | {
+        type: 'static';
+        startDate?: string;
+        endDate?: string;
+      }
+    | {
+        type: 'dynamic';
+        fromNow: string;
+      };
+  scenarioId?: string[];
+};
+
+export type DecisionFiltersWithPagination =
+  FiltersWithPagination<DecisionFilters>;
+
 export interface DecisionRepository {
-  listDecisions(args: {
-    outcome?: Outcome[];
-    triggerObject?: string[];
-    dateRange?:
-      | {
-          type: 'static';
-          startDate?: string;
-          endDate?: string;
-        }
-      | {
-          type: 'dynamic';
-          fromNow: string;
-        };
-    scenarioId?: string[];
-    offsetId?: string;
-    next?: boolean;
-    previous?: boolean;
-    limit?: number;
-    order?: 'ASC' | 'DESC';
-    sorting?: 'created_at';
-  }): Promise<PaginatedResponse<Decision>>;
+  listDecisions(
+    args: DecisionFiltersWithPagination,
+  ): Promise<PaginatedResponse<Decision>>;
 }
 
 export function getDecisionRepository() {
